switch automatically between 8 and 32 threads.

+++ modules/ftm/tests/test_dmThreads.py
@@ -1,5 +1,4 @@
 import dm_testbench
-import pytest
 
 """
 Class starts dmThread tests and compares with expected result.
@@ -7,18 +6,21 @@ Class starts dmThread tests and compares with expected result.
 class UnitTestDatamasterThreads(dm_testbench.DmTestbench):
 
   def run_dmThreads(self, count):
-    self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'reset', 'all'), [0])
-    self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'clear', '-f'), [0])
-    self.resetAllCpus()
-    scheduleFile = f'pps-all-threads-cpu0-{count}.dot'
-    self.generate_schedule(scheduleFile, count)
-    self.addSchedule(f'../{scheduleFile}')
-    self.assertIn(count, range(1,len(self.patternNames)+1), f'Number of threads is {count}, not in {range(1,len(self.patternNames)+1)}')
-    for i in range(count):
-      self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'Pattern0_' + chr(self.patternNames[i]), '-t', str(i)), [0])
-    self.startAndGetSubprocessOutput((self.binaryDmCmd, self.datamaster), [0], 13 + count)
-    self.checkRunningThreadsCmd()
-    self.deleteFile(scheduleFile)
+    if count <= self.threadQuantity:
+      self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'reset', 'all'), [0])
+      self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'clear', '-f'), [0])
+      self.resetAllCpus()
+      scheduleFile = f'pps-all-threads-cpu0-{count}.dot'
+      self.generate_schedule(scheduleFile, count)
+      self.addSchedule(f'../{scheduleFile}')
+      self.assertIn(count, range(1,len(self.patternNames)+1), f'Number of threads is {count}, not in {range(1,len(self.patternNames)+1)}')
+      for i in range(count):
+        self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'Pattern0_' + chr(self.patternNames[i]), '-t', str(i)), [0])
+      self.startAndGetSubprocessOutput((self.binaryDmCmd, self.datamaster), [0], 13 + count)
+      self.checkRunningThreadsCmd()
+      self.deleteFile(scheduleFile)
+    else:
+      self.assertGreater(count, self.threadQuantity)
 
   def test_dmThreads1(self):
     self.run_dmThreads(1)
@@ -44,121 +46,80 @@ class UnitTestDatamasterThreads(dm_testbench.DmTestbench):
   def test_dmThreads8(self):
     self.run_dmThreads(8)
 
-  @pytest.mark.thread32
   def test_dmThreads9(self):
     self.run_dmThreads(9)
 
-  @pytest.mark.thread32
   def test_dmThreads10(self):
     self.run_dmThreads(10)
 
-  @pytest.mark.thread32
   def test_dmThreads11(self):
     self.run_dmThreads(11)
 
-  @pytest.mark.thread32
   def test_dmThreads12(self):
     self.run_dmThreads(12)
 
-  @pytest.mark.thread32
   def test_dmThreads13(self):
     self.run_dmThreads(13)
 
-  @pytest.mark.thread32
   def test_dmThreads14(self):
     self.run_dmThreads(14)
 
-  @pytest.mark.thread32
   def test_dmThreads15(self):
     self.run_dmThreads(15)
 
-  @pytest.mark.thread32
   def test_dmThreads16(self):
     self.run_dmThreads(16)
 
-  @pytest.mark.thread32
   def test_dmThreads17(self):
     self.run_dmThreads(17)
 
-  @pytest.mark.thread32
   def test_dmThreads18(self):
     self.run_dmThreads(18)
 
-  @pytest.mark.thread32
   def test_dmThreads19(self):
     self.run_dmThreads(19)
 
-  @pytest.mark.thread32
   def test_dmThreads20(self):
     self.run_dmThreads(20)
 
-  @pytest.mark.thread32
   def test_dmThreads21(self):
     self.run_dmThreads(21)
 
-  @pytest.mark.thread32
   def test_dmThreads22(self):
     self.run_dmThreads(22)
 
-  @pytest.mark.thread32
   def test_dmThreads23(self):
     self.run_dmThreads(23)
 
-  @pytest.mark.thread32
   def test_dmThreads24(self):
     self.run_dmThreads(24)
 
-  @pytest.mark.thread32
   def test_dmThreads25(self):
     self.run_dmThreads(25)
 
-  @pytest.mark.thread32
   def test_dmThreads26(self):
     self.run_dmThreads(26)
 
-  @pytest.mark.thread32
   def test_dmThreads27(self):
     self.run_dmThreads(27)
 
-  @pytest.mark.thread32
   def test_dmThreads28(self):
     self.run_dmThreads(28)
 
-  @pytest.mark.thread32
   def test_dmThreads29(self):
     self.run_dmThreads(29)
 
-  @pytest.mark.thread32
   def test_dmThreads30(self):
     self.run_dmThreads(30)
 
-  @pytest.mark.thread32
   def test_dmThreads31(self):
     self.run_dmThreads(31)
 
-  @pytest.mark.thread32
   def test_dmThreads32(self):
     self.run_dmThreads(32)
 
   def test_dmAllThreads_Cpu0123(self):
-    count = 8
-    self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'reset', 'all'), [0])
-    self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'clear', '-f'), [0])
-    self.resetAllCpus()
-    self.addSchedule('pps-all-threads-cpu0.dot')
-    self.addSchedule('pps-all-threads-cpu1.dot')
-    self.addSchedule('pps-all-threads-cpu2.dot')
-    self.addSchedule('pps-all-threads-cpu3.dot')
-    index = 0
-    threadList = [('a', '0'), ('b', '1'), ('c', '2'), ('d', '3'), ('e', '4'), ('f', '5'), ('g', '6'), ('h', '7')]
-    for x, y in threadList:
-      if index < count:
-        self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'PPS0' + x, '-t', y), [0])
-        self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'PPS1' + x, '-t', y), [0])
-        self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'PPS2' + x, '-t', y), [0])
-        self.startAndCheckSubprocess((self.binaryDmCmd, self.datamaster, 'startpattern', 'PPS3' + x, '-t', y), [0])
-        index = index + 1
-    self.checkRunningThreadsCmd()
+    self.prepareRunThreads()
 
   patternNames = list(range(ord('a'), ord('z')+1)) + list(range(ord('A'), ord('Z')+1))
